third_party/WebKit/LayoutTests/fast/html/marquee-without-frame-no-crash.html

<script src="../../resources/js-test.js"></script>
foo
<script>
function handler() {
    shouldBeUndefined(window.open("foo", "bar"));
}
</script>
<template></template>
<iframe onload="handler()"></iframe>
<template>
<marquee>
